feat: Implement user authorization for profile management and update routes

+++ routes/web.php
@@ -78,9 +78,9 @@ Route::post('/password/reset', [ResetPasswordController::class, 'reset'])->name(
 
 // Profile routes
 Route::get('/profile', [UserController::class, 'showProfile'])->name('profile.show')->middleware('auth');
-Route::get('/profile/edit', [UserController::class, 'showEditProfile'])->name('profile.edit')->middleware('auth');
-Route::post('/profile/update', [UserController::class, 'update'])->name('profile.update')->middleware('auth');
-Route::delete('/profile/delete/{id}', [UserController::class, 'destroy'])->name('profile.delete')->middleware('auth');
+Route::get('/profile/edit/{id}', [UserController::class, 'showEditProfile'])->name('profile.edit')->middleware('auth');
+Route::post('/profile/update/{id}', [UserController::class, 'update'])->name('profile.update')->middleware('auth');
+Route::delete('/profile/delete/{id}', [UserController::class, 'destroy'])->name('profile.delete')->middleware('auth','admin');
 
 // Admin 
 Route::get('/admin', [AdminController::class, 'index'])->name('admin.index')->middleware('auth','admin');
